refactor(cloud): move parseApiErrorMessage into @deepnote/cloud
It was hoisted into @deepnote/database-integrations only so @deepnote/cloud
could share it, but it has no consumer there and nothing to do with database
integrations. Its only users are cloud-runs.ts and the CLI's import-client.ts,
and the CLI already depends on @deepnote/cloud — so cloud is the right home
with no new dependency edge or cycle. ApiError stays in database-integrations,
where fetch-integrations.ts still throws it.
